Pull out spec for choice function
