+++ curve25519-dalek/src/backend/serial/u64/scalar.rs
@@ -14,6 +14,7 @@
 use core::fmt::Debug;
 use core::ops::{Index, IndexMut};
 use subtle::{Choice, ConditionallySelectable};
+use super::choice_spec::*;
 
 #[cfg(feature = "zeroize")]
 use zeroize::Zeroize;
@@ -29,27 +30,6 @@ use vstd::calc;
 use vstd::prelude::*;
 
 verus! {
-
-// Tell Verus what Choice does
-
-#[verifier::external_type_specification]
-#[verifier::external_body]
-#[allow(dead_code)]
-pub struct ExChoice(Choice);
-
-pub uninterp spec fn boolify(c: Choice) -> bool;
-
-pub assume_specification [Choice::from](u: u8) -> (c: Choice)
-    ensures u == 0 ==> boolify(c) == false,
-            u == 1 ==> boolify(c) == true;
-
-#[verifier::external_body]
-fn select(x: &u64, y: &u64, c: Choice) -> (res: u64)
-    ensures boolify(c) ==> res == x,
-            ! boolify(c) ==> res == y
-{
-    u64::conditional_select(x, y, c)
-}
 
 
 // Spec functions
